Address comments for transducer
