+++ hyperion/torch/data/__init__.py
@@ -11,5 +11,5 @@ from .audio_dataset import AudioDataset
 
 # samplers
 from .bucketing_seg_sampler import BucketingSegSampler
-from .weighted_seq_sampler import ClassWeightedSeqSampler
+# from .weighted_seq_sampler import ClassWeightedSeqSampler
 from .seg_sampler_factory import SegSamplerFactory
